Moved package of selector and condition class
